+++ CHANGELOG.md
@@ -7,9 +7,8 @@
 ### Bug fixes
 
 ### Deprecated functionality
-* Deprecated `skbio.util.flatten`. Please use standard python library functionality
-described here (http://stackoverflow.com/questions/952914/making-a-flat-list-out-of-list-of-lists-in-python)
- (http://stackoverflow.com/questions/406121/flattening-a-shallow-list-in-python) ([#833](https://github.com/biocore/scikit-bio/issues/833))
+* Deprecated `skbio.util.flatten`. This function will be removed in scikit-bio 0.3.1. Please use standard python library functionality
+described here [Making a flat list out of lists of lists](http://stackoverflow.com/a/952952/3639023), [Flattening a shallow list](http://stackoverflow.com/a/406199/3639023) ([#833](https://github.com/biocore/scikit-bio/issues/833))
 
 ### Backward-incompatible changes
 * Removed the following deprecated functionality: